fixed assert to deal with no data in arch processor improved analyzer widget slightly greatly reduced exceptions produced during disassembly, this should speed things up a bit

+++ trunk/src/widgets/QDisassemblyView.h
@@ -79,7 +79,7 @@ private:
 	QString format_instruction_bytes(const edb::Instruction &insn, int maxStringPx, const QFontMetrics &metrics) const;
 	QString format_invalid_instruction_bytes(const edb::Instruction &insn, QPainter &painter) const;
 	edb::address_t address_from_coord(int x, int y) const;
-	size_t length_disasm_back(const quint8 *buf, int size) const;
+	size_t length_disasm_back(const quint8 *buf, size_t size) const;
 	edb::address_t previous_instructions(edb::address_t current_address, int count);
 	edb::address_t following_instructions(edb::address_t current_address, int count);
 	int address_length() const;
